UUID to Name, Character and User Session Contracts (#98)
* plan is in place, ready to start

* wip

* claude made an alembic revision with a same version number, now psql and alembic are pissed - removing the volume and renaming the version didnt fix it... WIP

* removed broken migration and re mirgrated with alembic this time

* still WIP but made a start on removing raw access to names, replacing with user uid

* realised the issue with local moderator state, touches the same issues in our sheleved mod ETL plan - this is a bigger slice of work will need to do in a separate branch

* claude thinks theres a workaround that needs no etl changes

* replaced two plans with one for service based calls on mod/role perm changes

* added schemas for setting role request/response and a command for setting

* fix map load bug still wanting usernames, and removed that log entirely, removed name mutations where were not using them anymore

* created a new contract to shape the dm attributes, and hydrate as part of the initial request so that we dont need to nest dm data inside player data

* higher level contract for session user and character data hyrdates through the old character contract, but under the session user contract so we can let users without characters join, i.e mods, spectators

* moderator request flow

* mod lock and state event to dashboard via WS is working, looks smooth!

* removed dead patterns

* reconciled plan files

* replaced dm chair fragment with a formal div to apply mb to the entire section, creating clearer separation between dm and mods, versus, player chars

* actioned PR feedback

---------

+++ rollplay-shared-contracts/shared_contracts/map.py
@@ -27,7 +27,7 @@ class GridConfig(ContractModel):
     colors: Optional[Dict[str, GridColorMode]] = None  # "edit_mode", "display_mode"
     offset_x: int = 0  # Whole-grid X shift (image px, can be negative)
     offset_y: int = 0  # Whole-grid Y shift (image px, can be negative)
-    grid_cell_size: Optional[int] = None  # Absolute cell size in native image pixels; None = not yet tuned
+    grid_cell_size: Optional[float] = None  # Absolute cell size in native image pixels; None = not yet tuned
 
 
 class MapConfig(ContractModel):
